+++ backend/scripts/map-gen/command-cluster.ts
@@ -1,7 +1,7 @@
 import { createRng, mulberry32Prng, type Rng } from "@guillaume-docquier/tools-ts"
 import { Command } from "commander"
-import { clusterGenerator } from "#lib/map/points/cluster.generator.ts"
-import type { Point2D } from "#lib/map/points/Point2D.ts"
+import { clusterGenerator } from "#lib/map-generation/points/cluster.generator.ts"
+import type { Point2D } from "#lib/map-generation/points/Point2D.ts"
 import { randomUInt32 } from "#lib/randomUInt32.ts"
 import { Parser } from "./Parser.ts"
 import { SvgRenderer } from "./SvgRenderer.ts"
