+++ backend/scripts/map-gen/command-galaxy.ts
@@ -1,8 +1,8 @@
 import path from "node:path"
 import { createRng, Distance, mulberry32Prng, UnitOfDistance, type Rng } from "@guillaume-docquier/tools-ts"
 import { Command } from "commander"
-import { galaxyGenerator } from "#lib/map/galaxy.generator.ts"
-import type { Point2D } from "#lib/map/points/Point2D.ts"
+import { galaxyGenerator } from "#lib/map-generation/galaxy.generator.ts"
+import type { Point2D } from "#lib/map-generation/points/Point2D.ts"
 import { createClusterCommand, type ClusterRenderOptions } from "./command-cluster.ts"
 import { createSpiralCommand, type SpiralRenderOptions } from "./command-spiral.ts"
 import { Parser } from "./Parser.ts"
