+++ backend/scripts/map-gen/command-spiral.ts
@@ -1,7 +1,7 @@
 import { createRng, mulberry32Prng, type Rng } from "@guillaume-docquier/tools-ts"
 import { Command } from "commander"
-import type { Point2D } from "#lib/map/points/Point2D.ts"
-import { spiralGenerator } from "#lib/map/points/spiral.generator.ts"
+import type { Point2D } from "#lib/map-generation/points/Point2D.ts"
+import { spiralGenerator } from "#lib/map-generation/points/spiral.generator.ts"
 import { randomUInt32 } from "#lib/randomUInt32.ts"
 import { Parser } from "./Parser.ts"
 import { SvgRenderer } from "./SvgRenderer.ts"
