+++ backend/src/lib/map-generation/points/Point2D.ts
